Update and rename CurrentDistribution.py to CurrentPath.py
cleanup and comments
